restructured component file system
