+++ public/images/landing-page/landing-page-main.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=about:internet
