+++ src/app/(authentication)/sign-up/page.tsx
@@ -1,7 +1,5 @@
-import Image from "next/image";
-import Link from "next/link";
 import styles from "./page.module.css";
-import Card from "../(components)/(card)/Card";
+import Card from "@/components/authentication/card/Card";
 import { getPageSession } from "@/lib/session";
 import { redirect } from "next/navigation";
