+++ src/app/(with-nav)/layout.tsx
@@ -1,10 +1,14 @@
 import type { Metadata } from "next";
-import { Inter, Montserrat } from "next/font/google";
 import "../globals.css";
 import "@fortawesome/fontawesome-free/css/all.min.css";
 import Navbar from "@/components/nav/Navbar";
-
+import { Inter, Montserrat, Poppins } from "next/font/google";
 const inter = Inter({ subsets: ["latin"] });
+const montserrat = Montserrat({ subsets: ["latin"] });
+const poppins = Poppins({
+    subsets: ["latin"],
+    weight: "300",
+});
 
 export const metadata: Metadata = {
     title: "My AI Resumes",
@@ -18,7 +22,7 @@ export default function RootLayout({
 }) {
     return (
         <html lang="en">
-            <body className={inter.className}>
+            <body className={poppins.className}>
                 <Navbar />
                 {children}
             </body>
