+++ src/components/authentication/email-verification-card/Card.tsx
