+++ src/components/authentication/sign-out-button/Button.tsx
