--- src/components/landing-page/title/Title.module.css
@@ -1,34 +1,0 @@
-.titleContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    gap: var(--gap-small);
-}
-
-.title {
-    font-size: var(--font-size-medium);
-    font-weight: bold;
-    color: white;
-    text-align: center;
-}
-
-.subtitle {
-    font-size: var(--font-size-x2large);
-    text-align: center;
-    font-weight: bold;
-    letter-spacing: 0.1em;
-}
-
-.description {
-    padding-block: var(--padding-block-medium);
-    font-size: var(--font-size-medium);
-    text-align: center;
-}
-
-.getStartedText {
-    font-size: var(--font-size-medium);
-    padding-block-start: var(--padding-block-medium);
-    color: white;
-}
