+++ src/components/landing-page/title/Title.tsx
@@ -2,18 +2,18 @@ import styles from "./Title.module.css";
 
 const Title = () => {
     return (
-        <div className={styles.titleContainer}>
+        <>
+            <p className={styles.description}>AI Resume Builder</p>
             <h1 className={styles.title}>
-                Artificial Intelligence Resume Builder
+                Effortlessly Craft Tailored Resumes for Every Job - Outshine
+                with Ease!
             </h1>
-            <h2 className={styles.subtitle}>
-                Create your perfect resume in <span>minutes</span> with the help
-                of <span>AI.</span>
-            </h2>
-            <p className={styles.getStartedText}>
-                Click on any section of the resume to enhance its content.
+            <p className={styles.subtitle}>
+                Harness the power of artificial intelligence to tailor your
+                resume uniquely for each job, significantly enhancing your
+                prospects of landing the job.
             </p>
-        </div>
+        </>
     );
 };
